+++ test/date.spec.js
@@ -1,16 +1,34 @@
 var testHelper = require('../src/testing/test-helper.js');
 var errorFormatter = testHelper.validationErrorFormatter;
 
-describe('Date validation type', function() {
+describe('Date validation type:', function() {
   beforeEach(function() {
     testHelper.initSyncFunction('build/sync-functions/test-date-sync-function.js');
   });
 
   describe('format validation', function() {
-    it('accepts a valid date', function() {
+    it('accepts a valid date with all components', function() {
       var doc = {
         _id: 'dateDoc',
         formatValidationProp: '2016-07-17'
+      };
+
+      testHelper.verifyDocumentCreated(doc);
+    });
+
+    it('accepts a valid date without a day', function() {
+      var doc = {
+        _id: 'dateDoc',
+        formatValidationProp: '0000-12'
+      };
+
+      testHelper.verifyDocumentCreated(doc);
+    });
+
+    it('accepts a valid date without a month or day', function() {
+      var doc = {
+        _id: 'dateDoc',
+        formatValidationProp: '9999'
       };
 
       testHelper.verifyDocumentCreated(doc);
@@ -43,10 +61,57 @@ describe('Date validation type', function() {
       testHelper.verifyDocumentNotCreated(doc, 'dateDoc', errorFormatter.dateFormatInvalid('formatValidationProp'));
     });
 
-    it('accepts a date that is within the minimum and maximum values', function() {
+    it('rejects a date with time and time zone components', function() {
       var doc = {
         _id: 'dateDoc',
-        rangeValidationProp: '2016-06-23'
+        formatValidationProp: '2016-07-17T15:01:58.382-05:00'
+      };
+
+      testHelper.verifyDocumentNotCreated(doc, 'dateDoc', errorFormatter.dateFormatInvalid('formatValidationProp'));
+    });
+
+    it('rejects a date with a time but no time zone', function() {
+      var doc = {
+        _id: 'dateDoc',
+        formatValidationProp: '2016-07-17T21:27:10.894'
+      };
+
+      testHelper.verifyDocumentNotCreated(doc, 'dateDoc', errorFormatter.dateFormatInvalid('formatValidationProp'));
+    });
+
+    it('rejects a value that is not a string', function() {
+      var doc = {
+        _id: 'dateDoc',
+        formatValidationProp: 20160717
+      };
+
+      testHelper.verifyDocumentNotCreated(doc, 'dateDoc', errorFormatter.typeConstraintViolation('formatValidationProp', 'date'));
+    });
+  });
+
+  describe('inclusive range constraints', function() {
+    it('accepts a date with all components that is within the minimum and maximum values', function() {
+      var doc = {
+        _id: 'dateDoc',
+        inclusiveRangeValidationProp: '2016-01-01'
+      };
+
+      testHelper.verifyDocumentCreated(doc);
+    });
+
+    it('accepts a date without a day that is within the minimum and maximum values', function() {
+      var doc = {
+        _id: 'dateDoc',
+        inclusiveRangeValidationProp: '2016-01'
+      };
+
+      testHelper.verifyDocumentCreated(doc);
+    });
+
+    it('accepts a date with a month or day that is within the minimum and maximum values', function() {
+      var doc = {
+        _id: 'dateDoc',
+        inclusiveRangeValidationProp: '2016'
       };
 
       testHelper.verifyDocumentCreated(doc);
@@ -55,25 +120,93 @@ describe('Date validation type', function() {
     it('rejects a date that is less than the minimum value', function() {
       var doc = {
         _id: 'dateDoc',
-        rangeValidationProp: '2016-06-22'
+        inclusiveRangeValidationProp: '2015-12-31'
       };
 
       testHelper.verifyDocumentNotCreated(
         doc,
         'dateDoc',
-        errorFormatter.minimumValueViolation('rangeValidationProp', '2016-06-23'));
+        errorFormatter.minimumValueViolation('inclusiveRangeValidationProp', '2015-12-31T23:59:59.999Z'));
     });
 
     it('rejects a date that is greater than the maximum value', function() {
       var doc = {
         _id: 'dateDoc',
-        rangeValidationProp: '2016-06-24'
+        inclusiveRangeValidationProp: '2016-01-02'
       };
 
       testHelper.verifyDocumentNotCreated(
         doc,
         'dateDoc',
-        errorFormatter.maximumValueViolation('rangeValidationProp', '2016-06-23T23:59:59.999Z'));
+        errorFormatter.maximumValueViolation('inclusiveRangeValidationProp', '2016-01-01T23:59:59.999Z'));
+    });
+  });
+
+  describe('exclusive range constraints', function() {
+    it('accepts a date with all components that is within the minimum and maximum values', function() {
+      var doc = {
+        _id: 'dateDoc',
+        exclusiveRangeValidationProp: '2018-02-01'
+      };
+
+      testHelper.verifyDocumentCreated(doc);
+    });
+
+    it('accepts a date without a day that is within the minimum and maximum values', function() {
+      var doc = {
+        _id: 'dateDoc',
+        exclusiveRangeValidationProp: '2018-02'
+      };
+
+      testHelper.verifyDocumentCreated(doc);
+    });
+
+    it('rejects a date that is less than the minimum value', function() {
+      var doc = {
+        _id: 'dateDoc',
+        exclusiveRangeValidationProp: '2017-12-31'
+      };
+
+      testHelper.verifyDocumentNotCreated(
+        doc,
+        'dateDoc',
+        errorFormatter.minimumValueExclusiveViolation('exclusiveRangeValidationProp', '2018'));
+    });
+
+    it('rejects a date that is equal to the minimum value', function() {
+      var doc = {
+        _id: 'dateDoc',
+        exclusiveRangeValidationProp: '2018'
+      };
+
+      testHelper.verifyDocumentNotCreated(
+        doc,
+        'dateDoc',
+        errorFormatter.minimumValueExclusiveViolation('exclusiveRangeValidationProp', '2018'));
+    });
+
+    it('rejects a date that is greater than the maximum value', function() {
+      var doc = {
+        _id: 'dateDoc',
+        exclusiveRangeValidationProp: '2018-02-03'
+      };
+
+      testHelper.verifyDocumentNotCreated(
+        doc,
+        'dateDoc',
+        errorFormatter.maximumValueExclusiveViolation('exclusiveRangeValidationProp', '2018-02-02'));
+    });
+
+    it('rejects a date that is equal to the maximum value', function() {
+      var doc = {
+        _id: 'dateDoc',
+        exclusiveRangeValidationProp: '2018-02-02'
+      };
+
+      testHelper.verifyDocumentNotCreated(
+        doc,
+        'dateDoc',
+        errorFormatter.maximumValueExclusiveViolation('exclusiveRangeValidationProp', '2018-02-02'));
     });
   });
 });